third_party/WebKit/LayoutTests/platform/win/http/tests/security/contentSecurityPolicy/require-sri-for/require-sri-for-sharedworker-allowed-expected.txt

CONSOLE ERROR: line 8: Refused to load the shared worker 'http://127.0.0.1:8000/security/contentSecurityPolicy/require-sri-for/sri-sharedworker.js' because 'require-sri-for' directive requires integrity attribute be present for all shared workers.
This is a testharness.js-based test.
PASS Test that a sharedworker can be created 
Harness: the test ran to completion.
